requirements.txt
telethon>=1.24.0
aiofiles>=23.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
